+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All important changes to the decitala package will be documented here.
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and the project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [v0.5.1] - January 17, 2020 (Kent, CT)
+### Added
+- Added a function in database.py that filters out cross-corpus duplicates from the data generated in ``create_database`` (``database.remove_cross_corpus_duplicates``)
+
 ### Changed
 - Fragment Trees are now created with ``frag_types`` via the class method ``FragmentTree.from_frag_type``. (The logic in the ``__init__`` is now much cleaner.)